+++ src/app/loadout-builder/LoadoutBuilder.tsx
@@ -1,7 +1,9 @@
-import { LoadoutParameters } from '@destinyitemmanager/dim-api-types';
 import { DestinyAccount } from 'app/accounts/destiny-account';
 import { createLoadoutShare } from 'app/dim-api/dim-api';
-import { savedLoadoutParametersSelector } from 'app/dim-api/selectors';
+import {
+  savedLoadoutParametersSelector,
+  savedLoStatConstraintsByClassSelector,
+} from 'app/dim-api/selectors';
 import CharacterSelect from 'app/dim-ui/CharacterSelect';
 import CollapsibleTitle from 'app/dim-ui/CollapsibleTitle';
 import PageWithMenu from 'app/dim-ui/PageWithMenu';
@@ -29,7 +31,7 @@ import { DestinyClass } from 'bungie-api-ts/destiny2';
 import { BucketHashes } from 'data/d2/generated-enums';
 import { AnimatePresence, motion } from 'framer-motion';
 import _ from 'lodash';
-import { memo, useEffect, useMemo } from 'react';
+import { memo, useCallback, useEffect, useMemo } from 'react';
 import { useSelector } from 'react-redux';
 import { createSelector } from 'reselect';
 import { allItemsSelector } from '../inventory/selectors';
@@ -45,6 +47,7 @@ import { filterItems } from './item-filter';
 import { useLbState } from './loadout-builder-reducer';
 import { buildLoadoutParams } from './loadout-params';
 import styles from './LoadoutBuilder.m.scss';
+import { LoUrlParams } from './LoadoutBuilderContainer';
 import { useProcess } from './process/useProcess';
 import {
   ArmorEnergyRules,
@@ -105,17 +108,13 @@ const halfTierModsSelector = createSelector(
 export default memo(function LoadoutBuilder({
   account,
   stores,
+  urlParameters,
   preloadedLoadout,
-  initialClassType,
-  notes,
-  initialLoadoutParameters,
 }: {
-  stores: DimStore[];
-  initialClassType: DestinyClass | undefined;
-  notes: string | undefined;
-  preloadedLoadout: Loadout | undefined;
-  initialLoadoutParameters: LoadoutParameters;
   account: DestinyAccount;
+  stores: DimStore[];
+  urlParameters: LoUrlParams | undefined;
+  preloadedLoadout: Loadout | undefined;
 }) {
   const defs = useD2Definitions()!;
   const allLoadouts = useSelector(loadoutsSelector);
@@ -124,6 +123,7 @@ export default memo(function LoadoutBuilder({
   const searchFilter = useSelector(searchFilterSelector);
   const searchQuery = useSelector(querySelector);
   const halfTierMods = useSelector(halfTierModsSelector);
+  const savedStatConstraintsByClass = useSelector(savedLoStatConstraintsByClassSelector);
 
   /** Gets items for the loadout builder and creates a mapping of classType -> bucketHash -> item array. */
   const items = useMemo(() => {
@@ -146,6 +146,8 @@ export default memo(function LoadoutBuilder({
     return items;
   }, [allItems]);
 
+  const optimizingLoadoutId = preloadedLoadout?.id;
+
   const [
     {
       loadoutParameters,
@@ -159,7 +161,7 @@ export default memo(function LoadoutBuilder({
       compareSet,
     },
     lbDispatch,
-  ] = useLbState(stores, preloadedLoadout, initialClassType, initialLoadoutParameters, defs);
+  ] = useLbState(stores, defs, preloadedLoadout, urlParameters);
   const isPhonePortrait = useIsPhonePortrait();
 
   const lockedExoticHash = loadoutParameters.exoticArmorHash;
@@ -172,35 +174,67 @@ export default memo(function LoadoutBuilder({
     [defs, loadoutParameters.mods]
   );
 
+  const selectedStore = stores.find((store) => store.id === selectedStoreId)!;
+  const classType = selectedStore.classType;
+
   // Save a subset of the loadout parameters to settings in order to remember them between sessions
   const setSetting = useSetSetting();
   useEffect(() => {
+    // If the user is playing with an existing loadout (potentially one they received from a loadout share)
+    // or a direct /optimizer link, do not overwrite the global saved loadout parameters.
+    // If they decide to save that loadout, these will still be saved with the loadout.
+    if (preloadedLoadout || urlParameters) {
+      return;
+    }
+
+    const newLoadoutParameters = buildLoadoutParams(
+      loadoutParameters,
+      '', // and the search query
+      // and don't save stat ranges either, just whether they're ignored
+      _.mapValues(statFilters, (m) => ({
+        ignored: m.ignored,
+        min: 0,
+        max: 10,
+      })),
+      statOrder
+    );
     const newSavedLoadoutParams = _.pick(
-      buildLoadoutParams(
-        loadoutParameters,
-        '', // and the search query
-        // and don't save stat ranges either, just whether they're ignored
-        _.mapValues(statFilters, (m) => ({
-          ignored: m.ignored,
-          min: 0,
-          max: 10,
-        })),
-        statOrder
-      ),
-      // Only keep a few parameters
-      'statConstraints',
+      newLoadoutParameters,
       'assumeArmorMasterwork',
       'lockArmorEnergyType'
     );
 
     setSetting('loParameters', newSavedLoadoutParams);
-  }, [setSetting, statFilters, statOrder, loadoutParameters]);
+    setSetting('loStatConstraintsByClass', {
+      ...savedStatConstraintsByClass,
+      [classType]: newLoadoutParameters.statConstraints,
+    });
+  }, [
+    setSetting,
+    statFilters,
+    statOrder,
+    loadoutParameters,
+    optimizingLoadoutId,
+    savedStatConstraintsByClass,
+    classType,
+    preloadedLoadout,
+    urlParameters,
+  ]);
+
+  const notes = urlParameters?.notes ?? preloadedLoadout?.notes;
+
+  const onCharacterChanged = useCallback(
+    (storeId: string) =>
+      lbDispatch({
+        type: 'changeCharacter',
+        store: stores.find((store) => store.id === storeId)!,
+        savedStatConstraintsByClass,
+      }),
+    [lbDispatch, savedStatConstraintsByClass, stores]
+  );
 
   // TODO: maybe load from URL state async and fire a dispatch?
   // TODO: save params to URL when they change? or leave it for the share...
-
-  const selectedStore = stores.find((store) => store.id === selectedStoreId)!;
-  const classType = selectedStore.classType;
 
   const enabledStats = useMemo(
     () => new Set(armorStats.filter((statType) => !statFilters[statType].ignored)),
@@ -225,7 +259,7 @@ export default memo(function LoadoutBuilder({
       ...loDefaultArmorEnergyRules,
       loadouts: {
         loadoutsByItem,
-        optimizingLoadoutId: preloadedLoadout?.id,
+        optimizingLoadoutId,
       },
     };
     if (loadoutParameters.lockArmorEnergyType !== undefined) {
@@ -249,7 +283,7 @@ export default memo(function LoadoutBuilder({
     loadoutParameters.lockArmorEnergyType,
     loadoutParameters.assumeArmorMasterwork,
     loadoutsByItem,
-    preloadedLoadout?.id,
+    optimizingLoadoutId,
     defs,
     characterItems,
     pinnedItems,
@@ -361,7 +395,7 @@ export default memo(function LoadoutBuilder({
         <CharacterSelect
           selectedStore={selectedStore}
           stores={stores}
-          onCharacterChanged={(storeId: string) => lbDispatch({ type: 'changeCharacter', storeId })}
+          onCharacterChanged={onCharacterChanged}
         />
         {isPhonePortrait ? (
           <CollapsibleTitle sectionId="lb-filter" title={t('LoadoutBuilder.Filter')}>
@@ -477,7 +511,7 @@ export default memo(function LoadoutBuilder({
               set={compareSet}
               selectedStore={selectedStore}
               loadouts={loadouts}
-              initialLoadoutId={preloadedLoadout?.id}
+              initialLoadoutId={optimizingLoadoutId}
               subclass={subclass}
               classType={classType}
               params={params}